TASK: Remove old `Uri` tests as the Flow Uri impl is gone
